Implement of a new command — remark
This command allows users of the AddressBook application to add optional remarks to people in their address book and edit it if required

Modify RemarkCommand to take in an Index and String for remark

Add RemarkCommandParser that will know how to parse two arguments, one
index and one with prefix 'r/'

Modify AddressBookParser to use RemarkCommandParser

Modify AddressBookParserTest to ensure that what the user input
generated the correct command.

Add Person#getRemark()

Modify PersonBuilder to support Remark.

Modify AddCommand to create a Person with a blank Remark.

Modify EditCommand to pass Remarks on unchanged.

Modify SampleDataUtil to add remarks for the sample data.

Add withRemark() to PersonBuilder.

Update tests and test data to use the remark field.

Replace the error message with the actual logic to modify the remarks
of a person.

Update RemarkCommandTest to test that execute() logic works.